Crate narwhal_consensus

Re-exports

pub use crate::consensus::Consensus;

Modules

bullshark

consensus

dag

metrics

tusk

Constants

DEFAULT_CHANNEL_SIZE

The default channel size used in the consensus and subscriber logic.

NUM_SHUTDOWN_RECEIVERS

The number of shutdown receivers to create on startup. We need one per component loop.